Vector2D & Vector3D is updated (double -> float)

+++ src/tools/vector2d.cpp
@@ -13,7 +13,7 @@ Vector2D::Vector2D(const Vector2D &vector)
     _x = vector._x; _y = vector._y;
 }
 
-Vector2D::Vector2D(double nx, double ny)
+Vector2D::Vector2D(float nx, float ny)
 {
     _x = nx;    _y = ny;
 }
@@ -44,18 +44,18 @@ void Vector2D::setZero()
     _x = _y = 0.0;
 }
 
-void Vector2D::rotate(double radian_rot)
+void Vector2D::rotate(float radian_rot)
 {
     _x = _x * cos(radian_rot) - _y * sin(radian_rot);
     _y = _x * sin(radian_rot) + _y * cos(radian_rot);
 }
 
-double Vector2D::lenght()
+float Vector2D::lenght()
 {
     return sqrt(_x*_x + _y*_y);
 }
 
-double Vector2D::arctan()
+float Vector2D::arctan()
 {
     return atan2(_y, _x);
 }
@@ -75,14 +75,14 @@ Vector2D Vector2D::operator -(const Vector2D &vector) const
     return Vector2D( _x - vector._x, _y - vector._y);
 }
 
-Vector2D Vector2D::operator *(double s)
+Vector2D Vector2D::operator *(float s)
 {
     return Vector2D( _x*s, _y*s);
 }
 
-Vector2D Vector2D::operator /(double s)
+Vector2D Vector2D::operator /(float s)
 {
-    double oneOverS = 1.0f / s;
+    float oneOverS = 1.0f / s;
     return Vector2D( _x*oneOverS, _y*oneOverS);
 }
 
@@ -98,24 +98,24 @@ Vector2D &Vector2D::operator -=(const Vector2D &vector)
     return *this;
 }
 
-Vector2D &Vector2D::operator *=(double s)
+Vector2D &Vector2D::operator *=(float s)
 {
     _x *= s; _y *= s;
     return *this;
 }
 
-Vector2D &Vector2D::operator /=(double s)
+Vector2D &Vector2D::operator /=(float s)
 {
-    double oneOverS = 1.0f / s;
+    float oneOverS = 1.0f / s;
     _x *= oneOverS; _y *= oneOverS;
     return *this;
 }
 
 void Vector2D::normalize()
 {
-    double magSqrd = _x*_x + _y*_y;
+    float magSqrd = _x*_x + _y*_y;
     if(magSqrd > 0.0f){
-        double oneOverMag = 1.0 / sqrt(magSqrd);
+        float oneOverMag = 1.0 / sqrt(magSqrd);
         _x *= oneOverMag;
         _y *= oneOverMag;
     }
@@ -124,10 +124,10 @@ void Vector2D::normalize()
 Vector2D Vector2D::normalized() const
 {
     Vector2D temp(this->_x, this->_y);
-    double magSqrd = _x*_x + _y*_y;
+    float magSqrd = _x*_x + _y*_y;
     if(magSqrd > 0.0f)
     {
-        double oneOverMag = 1.0 / sqrt(magSqrd);
+        float oneOverMag = 1.0 / sqrt(magSqrd);
         temp._x *= oneOverMag;
         temp._y *= oneOverMag;
     }
@@ -141,7 +141,7 @@ bool Vector2D::isInf() const
     return false;
 }
 
-double Vector2D::operator *(const Vector2D &vector)
+float Vector2D::operator *(const Vector2D &vector)
 {
     return (_x * vector._x) + (_y * vector._y);
 }
@@ -157,40 +157,40 @@ b2Vec2 Vector2D::b2vec2()
     return b2Vec2(this->_x, this->_y);
 }
 
-double Vector2D::vectorMag(const Vector2D &vector)
+float Vector2D::vectorMag(const Vector2D &vector)
 {
     return sqrt(vector._x * vector._x + vector._y * vector._y);
 }
 
-double Vector2D::distance(const Vector2D &a, const Vector2D &b)
+float Vector2D::distance(const Vector2D &a, const Vector2D &b)
 {
-    double dx = a._x - b._x;
-    double dy = a._y - b._y;
+    float dx = a._x - b._x;
+    float dy = a._y - b._y;
     return sqrt(dx * dx + dy * dy);
 }
 
-void Vector2D::set(double x_, double y_)
+void Vector2D::set(float x_, float y_)
 {
     this->_x = x_;
     this->_y = y_;
 }
 
-void Vector2D::setX(double x_)
+void Vector2D::setX(float x_)
 {
     this->_x = x_;
 }
 
-void Vector2D::setY(double y_)
+void Vector2D::setY(float y_)
 {
     this->_y = y_;
 }
 
-double Vector2D::X() const
+float Vector2D::X() const
 {
     return this->_x;
 }
 
-double Vector2D::Y() const
+float Vector2D::Y() const
 {
     return this->_y;
 }
