+++ src/tools/vector2d.h
@@ -6,18 +6,18 @@ class b2Vec2;
 
 class Vector2D
 {
-    double _x, _y;
+    float _x, _y;
 public:
-    void set(double x_, double y_);
-    void setX(double x_);
-    void setY(double y_);
+    void set(float x_, float y_);
+    void setX(float x_);
+    void setY(float y_);
 
-    double X() const;
-    double Y() const;
+    float X() const;
+    float Y() const;
 
     Vector2D();
     Vector2D(const Vector2D &vector);
-    Vector2D(double nx, double ny);
+    Vector2D(float nx, float ny);
     Vector2D(b2Vec2 vec);
 
     Vector2D operator =(const Vector2D &vector);
@@ -27,28 +27,28 @@ public:
     Vector2D normalized() const;
     bool isInf() const;
     void setZero();
-    void rotate(double radian_rot);
-    double lenght();
-    double arctan();
+    void rotate(float radian_rot);
+    float lenght();
+    float arctan();
 
     Vector2D operator -() const;
     Vector2D operator +(const Vector2D &vector) const;
     Vector2D operator -(const Vector2D &vector) const;
-    Vector2D operator *(double s);
-    Vector2D operator /(double s);
+    Vector2D operator *(float s);
+    Vector2D operator /(float s);
 
     Vector2D &operator +=(const Vector2D &vector);
     Vector2D &operator -=(const Vector2D &vector);
-    Vector2D &operator *=(double s);
-    Vector2D &operator /=(double s);
-    double operator *(const Vector2D &vector);
+    Vector2D &operator *=(float s);
+    Vector2D &operator /=(float s);
+    float operator *(const Vector2D &vector);
 
     Vector3D to3D();
     b2Vec2 b2vec2();
 
 //    Vector2D crossProduct(const Vector2D &a, const Vector2D &b);
-    static double vectorMag(const Vector2D  &vector);
-    static double distance(const Vector2D &a, const Vector2D &b);
+    static float vectorMag(const Vector2D  &vector);
+    static float distance(const Vector2D &a, const Vector2D &b);
 
 };
